% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_interface.R
\name{DataTypeReg}
\alias{DataTypeReg}
\title{R6 Class Representing a Data Type Registry}
\description{
Tba
}
\details{
Tba
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataTypeReg-get_template}{\code{DataTypeReg$get_template()}}
\item \href{#method-DataTypeReg-clone}{\code{DataTypeReg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTypeReg-get_template"></a>}}
\if{latex}{\out{\hypertarget{method-DataTypeReg-get_template}{}}}
\subsection{Method \code{get_template()}}{
The function used in any DTR to get a template
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypeReg$get_template(template_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template_id}}{The identifier of a template}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Not implemented, this is an interface
Not implemented, this is an interface
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTypeReg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataTypeReg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypeReg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
