% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_jsonld.R
\name{df_structure}
\alias{df_structure}
\title{Df structure function writes a dataframe into JSON-LD}
\usage{
df_structure(df, label)
}
\arguments{
\item{df}{A dataframe}

\item{label}{A character string}
}
\value{
An R object to be used by to_jsonld function
}
\description{
Df structure function writes a dataframe into JSON-LD
}
