% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_jsonld.R
\name{differ_type}
\alias{differ_type}
\title{An internal function to deal differently with inputs of different types}
\usage{
differ_type(input)
}
\arguments{
\item{input}{Can be NA, a dataframe, a tuple, or another input}
}
\value{
An R object to be used by to_jsonld function
}
\description{
An internal function to deal differently with inputs of different types
}
