% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtreg-package.R
\docType{package}
\name{dtreg-package}
\alias{dtreg}
\alias{dtreg-package}
\title{dtreg: Interact with Data Type Registries and Create Machine-Actionable Data}
\description{
You can load a schema from a data type registry (DTR) as an R object. Currently, the ePIC and the ORKG DTRs are supported. Use this schema to create an instance for reporting your research results. Write the instance in JSON-LD format to make it machine-actionable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OlgaLezhnina/dtreg}
  \item Report bugs at \url{https://github.com/OlgaLezhnina/dtreg/issues}
}

}
\author{
\strong{Maintainer}: Olga Lezhnina \email{olga.lezhnina@tib.eu} (\href{https://orcid.org/0000-0003-2237-7725}{ORCID})

Authors:
\itemize{
  \item Manuel Prinz \email{manuel.prinz@tib.eu} (\href{https://orcid.org/0000-0003-2151-4556}{ORCID})
  \item Markus Stocker \email{markus.stocker@tib.eu} (\href{https://orcid.org/0000-0001-5492-3212}{ORCID})
}

}
\keyword{internal}
