% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_interface.R
\name{dummy_r6}
\alias{dummy_r6}
\title{A dummy function explicitly referring to R6 package to declare the dependency}
\usage{
dummy_r6()
}
\description{
A dummy function explicitly referring to R6 package to declare the dependency
}
