% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_interface.R
\name{Epic}
\alias{Epic}
\title{Epic}
\arguments{
\item{schema_info}{An ePIC schema information}

\item{datatype_id}{The identifier of a datatype}

\item{prefix}{The URL prefix}

\item{identifier}{The ePIC property identifier}
}
\value{
A new object with the ePIC schema information

Extracted information from an ePIC schema

Context to include in JSON-LD file

Type to include in JSON-LD file

Type to include in JSON-LD file

Dataframe constants to include in JSON-LD file
}
\description{
R6 Class Representing the ePIC DTR

Create a new object with the schema information

Obtain information from an ePIC schema

Write ePIC-specific context for JSON-LD

Write ePIC-specific schema type for writing JSON-LD

Write ePIC-specific property type for writing JSON-LD

Write ePIC-specific dataframe constants for writing JSON-LD
}
\details{
The ePIC-specific information is provided by the dtr
}
\section{Fields}{

\describe{
\item{\code{schema_info}}{An ePIC schema information}
}}

\keyword{internal}
