% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_epic.R
\name{extract_epic}
\alias{extract_epic}
\title{Title}
\usage{
extract_epic(datatype_id)
}
\arguments{
\item{datatype_id}{The DOI of an ePIC schema}
}
\value{
An R object that contains information about the ePIC schema
}
\description{
Title
}
