% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_static.R
\name{from_static}
\alias{from_static}
\title{Title}
\usage{
from_static(template_doi)
}
\arguments{
\item{template_doi}{The DOI of a DTR template}
}
\value{
The template information from a JSON file, or "none" if not in static
}
\description{
Title
}
