% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_uid}
\alias{generate_uid}
\title{Generate uid function}
\usage{
generate_uid()
}
\value{
The counting function
}
\description{
Generate a counting function for assigning identifiers
}
\keyword{internal}
