% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_prefix}
\alias{get_prefix}
\title{Format a string}
\usage{
get_prefix(string)
}
\arguments{
\item{string}{A URL string}
}
\value{
The string which is the prefix of the URL
}
\description{
Format a string
}
