% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datatype.R
\name{load_datatype}
\alias{load_datatype}
\title{Load datatype}
\usage{
load_datatype(datatype_id)
}
\arguments{
\item{datatype_id}{The identifier of a datatype schema, such as URL}
}
\value{
a list of schemata as R objects
}
\description{
Load R objects with datatype schema information for creating a new instance
}
\examples{
dt <- dtreg::load_datatype("https://doi.org/21.T11969/1ea0e148d9bbe08335cd")

}
