% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_interface.R
\name{Orkg}
\alias{Orkg}
\title{ORKG}
\arguments{
\item{schema_info}{An ORKG template information}

\item{datatype_id}{The identifier of an ORKG template}

\item{prefix}{The URL prefix}

\item{identifier}{The ORKG property identifier}
}
\value{
A new object with the ORKG template information

Extracted information from an ORKG template

Context to include in JSON-LD file

Type to include in JSON-LD file

Type to include in JSON-LD file

Constants to include in JSON-LD file
}
\description{
R6 Class Representing the ORKG DTR

Create a new object with the schema information

Obtain information from an ORKG template

Write ORKG-specific context for JSON-LD

Write ORKG-specific template type for writing JSON-LD

Write ORKG-specific property type for writing JSON-LD

Write ORKG-specific dataframe constants for writing JSON-LD
}
\details{
The ORKG-specific information is provided by the dtr
}
\section{Fields}{

\describe{
\item{\code{schema_info}}{An ORKG template information}
}}

\keyword{internal}
