% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{range_split}
\alias{range_split}
\title{Split an ePIC cardinality string into min and max values}
\usage{
range_split(range_str)
}
\arguments{
\item{range_str}{An ePIC string for cardinality range}
}
\value{
A named list with min and max values as strings
}
\description{
Split an ePIC cardinality string into min and max values
}
