% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_epic.R
\name{request_epic}
\alias{request_epic}
\title{Title}
\usage{
request_epic(template_doi)
}
\arguments{
\item{template_doi}{The DOI of an ePIC template}
}
\value{
Requested information about an ePIC template
}
\description{
Title
}
