% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_interface.R
\name{select_dtr}
\alias{select_dtr}
\title{Select datatype registry}
\usage{
select_dtr(datatype_id)
}
\arguments{
\item{datatype_id}{The identifier of a datatype}
}
\value{
An R6 class for the specific datatype registry
}
\description{
Select datatype registry
}
