% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_fields}
\alias{show_fields}
\title{Title}
\usage{
show_fields(object)
}
\arguments{
\item{object}{An object from dtreg::load_objects}
}
\value{
Fields to use in an instance
}
\description{
Title
}
\examples{
pd <- load_objects("https://doi.org/21.T11969/1ea0e148d9bbe08335cd")
show_fields(pd$pidinst_schemaobject())

}
