% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{specify_cardinality}
\alias{specify_cardinality}
\title{Specify cardinality of an ePIC property}
\usage{
specify_cardinality(cardinality_string)
}
\arguments{
\item{cardinality_string}{An ePIC string for cardinality}
}
\value{
A named list with min and max values as integers or NA
}
\description{
Specify cardinality of an ePIC property
}
