% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datatype.R
\name{write_proxies}
\alias{write_proxies}
\title{Write proxies}
\usage{
write_proxies(r6_classes)
}
\arguments{
\item{r6_classes}{A list of R6 classes for a specified datatype}
}
\value{
a list of objects for creating new instances
}
\description{
Write proxies
}
