% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datatype.R
\name{write_r6_classes}
\alias{write_r6_classes}
\title{Title}
\usage{
write_r6_classes(template_doi)
}
\arguments{
\item{template_doi}{The DOI of a DTR template}
}
\value{
an R6 object for the template
}
\description{
Title
}
