% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_epic.R
\name{check_static}
\alias{check_static}
\title{Title}
\usage{
check_static(epic_id)
}
\arguments{
\item{epic_id}{The identifier of an ePIC object}
}
\value{
Boolean, whether the file is in the extdata
}
\description{
Title
}
