% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_static.R
\name{check_static}
\alias{check_static}
\title{Title}
\usage{
check_static(template_doi)
}
\arguments{
\item{template_doi}{The DOI of a DTR template}
}
\value{
The name of JSON file in the static data, or "none" if absent
}
\description{
Title
}
