% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_interface.R
\name{DataTypeReg}
\alias{DataTypeReg}
\title{DatatypReg}
\arguments{
\item{datatype_id}{The identifier of a datatype}

\item{prefix}{The URL prefix}

\item{identifier}{The property identifier}
}
\value{
Not implemented, this is an interface
Not implemented, this is an interface

Not implemented, this is an interface
Not implemented, this is an interface

Not implemented, this is an interface
Not implemented, this is an interface

Not implemented, this is an interface
Not implemented, this is an interface

Not implemented, this is an interface
Not implemented, this is an interface
}
\description{
Interface for R6 class representing a datatype registry

Obtain information from a datatype schema

Write dtr-specific context for JSON-LD

Write schema type for JSON-LD

Write property type for JSON-LD

Write dataframe constants for JSON-LD
}
\details{
The dtr-specific information is provided by the dtr
}
\keyword{internal}
