% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_interface.R
\name{DataTypeReg}
\alias{DataTypeReg}
\title{R6 Class Representing a Data Type Registry}
\description{
Tba
}
\details{
Tba
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataTypeReg-get_schema_info}{\code{DataTypeReg$get_schema_info()}}
\item \href{#method-DataTypeReg-add_context}{\code{DataTypeReg$add_context()}}
\item \href{#method-DataTypeReg-add_dt_type}{\code{DataTypeReg$add_dt_type()}}
\item \href{#method-DataTypeReg-add_dtp_type}{\code{DataTypeReg$add_dtp_type()}}
\item \href{#method-DataTypeReg-add_df_constants}{\code{DataTypeReg$add_df_constants()}}
\item \href{#method-DataTypeReg-clone}{\code{DataTypeReg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTypeReg-get_schema_info"></a>}}
\if{latex}{\out{\hypertarget{method-DataTypeReg-get_schema_info}{}}}
\subsection{Method \code{get_schema_info()}}{
The function used in any DTR to get a template
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypeReg$get_schema_info(datatype_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datatype_id}}{The identifier of a datatype}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Not implemented, this is an interface
Not implemented, this is an interface
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTypeReg-add_context"></a>}}
\if{latex}{\out{\hypertarget{method-DataTypeReg-add_context}{}}}
\subsection{Method \code{add_context()}}{
The function to write context for JSON-LD
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypeReg$add_context(prefix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{The URL prefix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Not implemented, this is an interface
Not implemented, this is an interface
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTypeReg-add_dt_type"></a>}}
\if{latex}{\out{\hypertarget{method-DataTypeReg-add_dt_type}{}}}
\subsection{Method \code{add_dt_type()}}{
The function to write schema type for JSON-LD
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypeReg$add_dt_type(identifier)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{The schema identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Not implemented, this is an interface
Not implemented, this is an interface
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTypeReg-add_dtp_type"></a>}}
\if{latex}{\out{\hypertarget{method-DataTypeReg-add_dtp_type}{}}}
\subsection{Method \code{add_dtp_type()}}{
The function to write property for JSON-LD
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypeReg$add_dtp_type(identifier)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{The property identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Not implemented, this is an interface
Not implemented, this is an interface
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTypeReg-add_df_constants"></a>}}
\if{latex}{\out{\hypertarget{method-DataTypeReg-add_df_constants}{}}}
\subsection{Method \code{add_df_constants()}}{
The function to write dataframe constants for JSON-LD
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypeReg$add_df_constants()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Not implemented, this is an interface
Not implemented, this is an interface
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTypeReg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataTypeReg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypeReg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
