% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_jsonld.R
\name{df_structure}
\alias{df_structure}
\title{Dataframe structure}
\usage{
df_structure(df, label)
}
\arguments{
\item{df}{A dataframe}

\item{label}{A string}
}
\value{
A named list to be used by to_jsonld function
}
\description{
Prepare a dataframe or a tuple for to_jsonld function
}
