% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_jsonld.R
\name{differ_input}
\alias{differ_input}
\title{Differ input deals differently with inputs of different types}
\usage{
differ_input(input)
}
\arguments{
\item{input}{Can be NA, a dataframe, a tuple, or another input}
}
\value{
An input-dependent R object or the df_structure function call
to be further used by to_jsonld function
}
\description{
Differ input deals differently with inputs of different types
}
