% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_jsonld.R
\name{differ_input}
\alias{differ_input}
\title{Differ input}
\usage{
differ_input(input)
}
\arguments{
\item{input}{A dataframe, a tuple, or another object}
}
\value{
The result of the df_structure function, or the input unchanged
}
\description{
Differentiate input for further use by to_jsonld function
}
