% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_epic.R
\name{extract_epic}
\alias{extract_epic}
\title{Title}
\usage{
extract_epic(template_doi)
}
\arguments{
\item{template_doi}{The DOI of an ePIC template}
}
\value{
An R object that contains information about the ePIC template
}
\description{
Title
}
