% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_epic.R
\name{extract_epic}
\alias{extract_epic}
\title{Title}
\usage{
extract_epic(epic_id)
}
\arguments{
\item{epic_id}{The identifier of an ePIC object}
}
\value{
An R object that contains information about the ePIC object
}
\description{
Title
}
