% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_orkg.R
\name{extract_orkg}
\alias{extract_orkg}
\title{Title}
\usage{
extract_orkg(dt_id)
}
\arguments{
\item{dt_id}{The URL of an ORKG template}
}
\value{
An R object that contains information about the ORKG template
}
\description{
Title
}
