% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_static.R
\name{from_static}
\alias{from_static}
\title{From static}
\usage{
from_static(datatype_id)
}
\arguments{
\item{datatype_id}{The identifier of a datatype, such as URL}
}
\value{
The R object with schema from a JSON file, or NULL if not in static
}
\description{
Get schema information from static files, or NULL if not in static
}
\keyword{internal}
