% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_static.R
\name{from_static}
\alias{from_static}
\title{From static, to get information from static files}
\usage{
from_static(datatype_id)
}
\arguments{
\item{datatype_id}{The identifier, such as URL, of a datatype}
}
\value{
The R object with schema from a JSON file, or NULL if not in static
}
\description{
From static, to get information from static files
}
