% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_prefix}
\alias{get_prefix}
\title{Get prefix}
\usage{
get_prefix(url_string)
}
\arguments{
\item{url_string}{A URL string}
}
\value{
The prefix string indicating a datatype
}
\description{
Get prefix of a URL string
}
\keyword{internal}
