% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datatype.R
\name{load_datatype}
\alias{load_datatype}
\title{Title}
\usage{
load_datatype(template_doi)
}
\arguments{
\item{template_doi}{The DOI of a DTR template}
}
\value{
a list of objects for creating new instances
}
\description{
Title
}
\examples{
pd <- load_datatype("https://doi.org/21.T11969/1ea0e148d9bbe08335cd")

}
