% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_interface.R
\name{Orkg}
\alias{Orkg}
\title{R6 Class Representing the ORKG DTR}
\description{
Tba
}
\details{
Tba
}
\section{Super class}{
\code{\link[dtreg:DataTypeReg]{dtreg::DataTypeReg}} -> \code{Orkg}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{template_info}}{An ORKG template information}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Orkg-new}{\code{Orkg$new()}}
\item \href{#method-Orkg-get_template_info}{\code{Orkg$get_template_info()}}
\item \href{#method-Orkg-add_context}{\code{Orkg$add_context()}}
\item \href{#method-Orkg-add_dt_type}{\code{Orkg$add_dt_type()}}
\item \href{#method-Orkg-clone}{\code{Orkg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orkg-new"></a>}}
\if{latex}{\out{\hypertarget{method-Orkg-new}{}}}
\subsection{Method \code{new()}}{
Create a new object with the template information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orkg$new(template_info = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template_info}}{An ORKG template information}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new object with the template information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orkg-get_template_info"></a>}}
\if{latex}{\out{\hypertarget{method-Orkg-get_template_info}{}}}
\subsection{Method \code{get_template_info()}}{
Write information from an ORKG template
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orkg$get_template_info(template_doi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template_doi}}{The URL of an ORKG template}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Extracted information from an ORKG template
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orkg-add_context"></a>}}
\if{latex}{\out{\hypertarget{method-Orkg-add_context}{}}}
\subsection{Method \code{add_context()}}{
Give ORKG-specific context for writing JSON-LD
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orkg$add_context(prefix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{The URL prefix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Context to include in JSON-LD file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orkg-add_dt_type"></a>}}
\if{latex}{\out{\hypertarget{method-Orkg-add_dt_type}{}}}
\subsection{Method \code{add_dt_type()}}{
Give ORKG-specific schema type for writing JSON-LD
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orkg$add_dt_type(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The schema identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Type to include in JSON-LD file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Orkg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Orkg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Orkg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
