% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_dtr.R
\name{request_dtr}
\alias{request_dtr}
\title{Request dtr}
\usage{
request_dtr(route)
}
\arguments{
\item{route}{The path for requesting a dtr API}
}
\value{
Requested information about a datatype schema
}
\description{
Request an API of a datatype registry
}
\keyword{internal}
