% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_epic.R
\name{request_epic}
\alias{request_epic}
\title{Title}
\usage{
request_epic(epic_id)
}
\arguments{
\item{epic_id}{The identifier of an ePIC object}
}
\value{
Requested information about an ePIC object
}
\description{
Title
}
