% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_interface.R
\name{select_dtr}
\alias{select_dtr}
\title{Title}
\usage{
select_dtr(template_doi)
}
\arguments{
\item{template_doi}{The DOI of a template in any DTR}
}
\value{
An R6 object for the specific DTR
}
\description{
Title
}
