% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_fields}
\alias{show_fields}
\title{Show fields that can be filled by the user when writing a datatype instance}
\usage{
show_fields(datatype)
}
\arguments{
\item{datatype}{An R object from dtreg::load_datatype}
}
\value{
Fields that can be used in an instance
}
\description{
Show fields that can be filled by the user when writing a datatype instance
}
\examples{
pd <- load_datatype("https://doi.org/21.T11969/1ea0e148d9bbe08335cd")
show_fields(pd$pidinst_schemaobject())

}
