% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_fields}
\alias{show_fields}
\title{Show fields}
\usage{
show_fields(datatype)
}
\arguments{
\item{datatype}{An R object from dtreg::load_datatype}
}
\value{
Fields that can be used in an instance
}
\description{
Show fields of a schema that can be used for creating an instance
}
\examples{
dt <- dtreg::load_datatype("https://doi.org/21.T11969/74bc7748b8cd520908bc")
dtreg::show_fields(dt$inferential_test_output())

}
