% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{specify_cardinality}
\alias{specify_cardinality}
\title{Specify cardinality}
\usage{
specify_cardinality(cardinality_string)
}
\arguments{
\item{cardinality_string}{An ePIC string for cardinality}
}
\value{
A named list with min and max values as integers or NA
}
\description{
Write cardinality of an ePIC property as a named list
}
