% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_jsonld.R
\name{to_jsonld}
\alias{to_jsonld}
\title{Write JSON-LD}
\usage{
to_jsonld(instance)
}
\arguments{
\item{instance}{An instance of a datatype-related R6 class}
}
\value{
JSON string in JSON-LD format
}
\description{
Write an instance into JSON-LD file
}
\examples{
dt <- dtreg::load_datatype("https://doi.org/21.T11969/74bc7748b8cd520908bc")
instance <- dt$inferential_test_output(label = "my_results")
result <- dtreg::to_jsonld(instance)

}
