% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_jsonld.R
\name{to_jsonld}
\alias{to_jsonld}
\title{To_jsonld writes an instance as JSON-LD}
\usage{
to_jsonld(instance)
}
\arguments{
\item{instance}{An instance of an R6 class}
}
\value{
JSON string in JSON-LD format
}
\description{
To_jsonld writes an instance as JSON-LD
}
