% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{write_proxies}
\alias{write_proxies}
\title{Title}
\usage{
write_proxies(object_list)
}
\arguments{
\item{object_list}{A list of R6 objects}
}
\value{
a list of objects for creating new instances
}
\description{
Title
}
