% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datatype.R
\name{write_r6_classes}
\alias{write_r6_classes}
\title{Write R6 classes for a specified datatype}
\usage{
write_r6_classes(datatype_id)
}
\arguments{
\item{datatype_id}{The identifier of a datatype schema, such as URL}
}
\value{
a list of R6 classes for the datatype
}
\description{
Write R6 classes for a specified datatype
}
