% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{write_r6_classes}
\alias{write_r6_classes}
\title{Title}
\usage{
write_r6_classes(templ_info)
}
\arguments{
\item{templ_info}{an R object with a template structured information}
}
\value{
an R6 object for the template
}
\description{
Title
}
